+++ packages/rma-server/src/warranty-claim/aggregates/warranty-claim-aggregate/warranty-claim-aggregate.service.ts
@@ -729,6 +729,13 @@ export class WarrantyClaimAggregateService extends AggregateRoot {
           }
           return of({});
         }),
+        switchMap(res => {
+          return from(
+            this.serialNoHistoryService.deleteOne({
+              parent_document: cancelPayload.uuid,
+            }),
+          );
+        }),
       );
   }
 }