+++ packages/rma-warranty/src/app/warranty-ui/view-warranty-claims/service-invoices/add-service-invoice/add-service-invoice.service.spec.ts
@@ -23,7 +23,14 @@ describe('AddServiceInvoiceService', () => {
         NoopAnimationsModule,
         RouterTestingModule.withRoutes([]),
       ],
-      providers: [{ provide: STORAGE_TOKEN, useValue: {} }],
+      providers: [
+        {
+          provide: STORAGE_TOKEN,
+          useValue: {
+            getItem: (...args) => Promise.resolve(),
+          },
+        },
+      ],
     });
     service = TestBed.inject(AddServiceInvoiceService);
   });