+++ packages/rma-warranty/src/app/warranty-ui/view-warranty-claims/service-invoices/service-invoices.component.spec.ts
@@ -32,7 +32,6 @@ describe('ServiceInvoicesComponent', () => {
           provide: AddServiceInvoiceService,
           useValue: {
             getServiceInvoiceList: (...args) => of([{}]),
-            syncInvoice: (...args) => of({}),
           },
         },
         {